🐞 Fix nested measured() blocks
If one `measured()` block is nested within another, then only
the outer block's callbacks would be invoked. This was happening
because the outer block's performance observer is notified of the
inner block's "mark" performance entry, but there was no validation
of whether the performance entry matches the one this particular
observer is expecting. **Any** performance entry of type "mark"
would have triggered all pending `measured()` blocks to complete,
in the order in which they were created.

This commit addresses this by validating the name of the performance
entry against the expected id for which that `measured()` block is
listening.

Additional test coverage has been added for this case.
The test script has now been modified to enforce strict 100% coverage
not only for lines, but also for functions, statements, and branches.
An additional test has been added to cover the optional `id` option,
which allows the caller to override the identifier used for peformance
entries.